Add rtl::unique_id(), include src/test/ in clang-tidy coverage
rtl::unique_id(thread): a 64-bit, monotonically increasing,
Twitter-snowflake-style id (41-bit ms timestamp | 10-bit thread |
12-bit sequence), thread-safe via one atomic + CAS retry loop. Needed
by ach's upcoming engine_t::open/closed/closing (eng_state.id is a
snowflake, generated by the caller, not the database).

src/test/ was excluded from the `tidy` custom target's -source-filter
(and unit_tests' own CXX_CLANG_TIDY property) because Catch2's macros
were assumed to trigger clang-tidy false positives - the same reasoning
ach's own .clang-tidy already worked around by turning off
cppcoreguidelines-avoid-do-while project-wide rather than skipping
files. Applying the same fix here and turning src/test/ coverage on
found ~60 real findings (not false positives) across the existing test
suite and rtl/db2/query.hpp, rtl/psql/query.hpp, async_db.hpp - fixed:

- readability-magic-numbers: named constants (test_t1.cpp, test_async.cpp)
- readability-function-cognitive-complexity: extracted helper functions/
  split into their own TEST_CASE (test_t1.cpp, test_async.cpp)
- cppcoreguidelines-pro-bounds-avoid-unchecked-container-access /
  -pro-bounds-pointer-arithmetic: NOLINT on the pre-checked span/vector
  indexing generated buffer code relies on (db2/query.hpp, psql/query.hpp,
  async_db.hpp, test_query_move.cpp - the last one via .at() instead)
  performance-no-int-to-ptr: NOLINT on the two SQLSetStmtAttr calls the
  ODBC API itself requires a size encoded as a pointer for (db2/query.hpp)
- misc-const-correctness: db2/query.hpp (conn/ret can be const),
  psql/query.hpp (holders - NOLINT, modified only in an if constexpr
  branch a has_p == false instantiation never compiles)
- readability-use-concise-preprocessor-directives, readability-container-
  contains, abseil-string-find-str-contains: db_error_test.cpp
- bugprone-empty-catch: NOLINT, deliberate fallback (test_db.hpp)
- bugprone-command-processor: NOLINT, the generator has no library API
  (test_parallel.cpp)
- misc-const-correctness, hicpp-invalid-access-moved (added to an
  existing NOLINT list): test_parallel.cpp, test_query_move.cpp
- cppcoreguidelines-narrowing-conversions: NOLINT, psql's count(*) is
  int64_t vs db2's int32_t, both compiled against the same test source
  (test_perf.cpp)
- readability-inconsistent-ifelse-braces: test_bench_async.cpp

+++ src/test/test_perf.cpp
@@ -82,7 +82,7 @@ namespace
     constexpr int first_printable = 0x20;
     constexpr int last_printable  = 0x7E;
 
-    std::mt19937                      gen(static_cast<uint32_t>(id));
+    std::mt19937                       gen(static_cast<uint32_t>(id));
     std::uniform_int_distribution<int> dist(first_printable, last_printable);
 
     std::string s(tran_width, '\0');
@@ -136,7 +136,12 @@ namespace
     auto got = cnt.fetch();
     require_ok(got, "fetch(perf_count)");
     REQUIRE(*got);
-    return cnt.get_result()->cnt();
+    // cnt() is psql's int64_t / db2's int32_t (count(*) is bigint on one
+    // backend, integer on the other) - this test never writes anywhere near
+    // INT32_MAX rows, and every caller already compares against an int32_t
+    // (perf_rows), so the cast is intentional (and a no-op on db2).
+    // NOLINTNEXTLINE(cppcoreguidelines-narrowing-conversions,readability-redundant-casting)
+    return static_cast<int32_t>(cnt.get_result()->cnt());
   }
 
   /// one row of a fetched window, checked against what write_all_rows put there
